taco/requirements.txt
PyQt6>=6.5.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
numpy>=1.24.0
pyrr>=0.10.3
